Fixed insets in personalInfo
