+++ src/lib/components/personalInfo/contributors.json
