+++ src/lib/components/recentResults/recents.svelte
@@ -2,6 +2,7 @@
     import { gatherRecentGrades } from "./gatherRecentGrades"
     import { gatherNotifications } from "$components/notifications/notifications";
     import RecentsDisplay from "./recentsDisplay.svelte";
+	import Card from "$components/degreeCalculator/card.svelte";
 
 
     const recentItemsFromCache: any[] = getCachedRecentItems();
