--- src/routes/faq/+page.svelte
@@ -1,7 +1,0 @@
-<script>
-	import FAQ from '$components/faq/faq.svelte';
-	import SubPageHeader from '$shared/subPageHeader.svelte';
-</script>
-
-<SubPageHeader title="FAQ" />
-<FAQ />
